Refactor Gemini provider into dedicated package
