+++ projects/04-llm-adapter-shadow/src/llm_adapter/providers/gemini/__init__.py
@@ -1,0 +1,6 @@
+"""Gemini provider package."""
+
+from .messages import parse_gemini_messages
+from .provider import GeminiProvider
+
+__all__ = ["GeminiProvider", "parse_gemini_messages"]
